Restructured the commands
Restructured the commands to make them cleaner and extendable.
