+++ src/Commands/CommandInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace willitscale\Streetlamp\Commands;
+
+interface CommandInterface
+{
+    public function command(?array $arguments = []): void;
+}
