+++ streetlamp
@@ -3,6 +3,15 @@
 
 require_once('vendor/autoload.php');
 
+use DI\Container;
 use willitscale\Streetlamp\Streetlamp;
 
-new Streetlamp($argc, $argv);
+try {
+    array_shift($argv);
+    (new Container())
+        ->make(Streetlamp::class)
+        ->command($argv);
+} catch (Exception $e) {
+    var_dump($e->getMessage());
+    exit(1);
+}
