+++ templates/apache.conf.tmpl
@@ -1,0 +1,15 @@
+<VirtualHost *:80>
+  ServerName localhost
+  DocumentRoot /app
+  
+  ErrorLog /dev/stderr
+  CustomLog /dev/stdout combined
+    
+  <IfModule mod_rewrite.c>
+    RewriteEngine on
+    RewriteCond %{REQUEST_FILENAME} !-f
+    RewriteCond %{REQUEST_FILENAME} !-d
+    RewriteRule ^(.*)$ index.php/$1 [L,QSA]
+  </IfModule>
+
+</VirtualHost>
