+++ templates/nginx.conf.tmpl
@@ -7,11 +7,8 @@ server {
   root /app;
   index index.php;
 
-  location ~* \.(js|ico|gif|jpg|png|svg|css|jpeg|eot|woff|ttf)$ {
-  }
-
-  location ~* ^/(?!(index\.php)) {
-    rewrite ^ /index.php$is_args$args;
+  location / {
+    try_files $uri $uri/ /index.php;
   }
 
   location ~ \.php$ {
